+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 
 This document keeps track of changes between releases of the library.
 
+v0.6.6
+------
+* ADDED: Support for listening/manipulating property collection on load
+* ADDED: Option to only run exclusion strategies by annotation
+
 v0.6.5
 ------
 * ADDED: PHPStorm metadata to help with auto completion